Bug fix -x option
It should not take an arg but it was specified to take an arg. This
resulted in in some cases the wrong number of globs being detected and
in other cases it reported that no glob was specified (even when there
was). It also fixes some other issues.

Updated disclaimer a bit about possible (but currently no known) bugs.