Add dry-run option -n and other things
The -n option allows one to see what would be catted if the -n option
was not used. Depending on verbosity it will show either just the file
names or the cat command (with or without options depending on if the -o
option was used) along with the file names.

Move check for git repo before doing anything else so the user need not
try and fix the command line only to realise they're not in a git repo.

Improve (as I see it at least) the order of the error messages that are
shown with the usage message. If one does not specify a glob, for
example, show the usage and then the error message. This way the error
stands out.

Possibly other script changes as well.

Updated README.md with fixes and other new examples (the -n option).